Changed many URIs to protocol relative URIs
The page would not load properly on my browser (which enforces HTTPS
using https://www.eff.org/https-everywhere).

With these changes many of the URIs now follow the protocol that was
used when accessing the page.

I only changed the URIs for href in ‘a’ and ‘link’.

I did not change the URIs in <span resource=“”> As I am unsure wether
the URI standard allows that and if it is a desired output.